Moved auth.ts to more appropriate folder. Updated imports
